requirements.txt
google-adk
python-dotenv
litellm
yfinance